refactor S3FileFilter class
