+++ cdap-app-templates/cdap-etl/cdap-etl-batch/pom.xml
@@ -62,6 +62,7 @@
       <scope>provided</scope>
     </dependency>
     <dependency>
+      <!-- change to hadoop-aws when hadoop dependency is updated to 2.6 -->
       <groupId>org.apache.hadoop</groupId>
       <artifactId>hadoop-common</artifactId>
       <type>test-jar</type>
