+++ cdap-app-templates/cdap-etl/cdap-etl-batch/src/test/java/org/apache/hadoop/fs/s3native/S3NInMemoryFileSystem.java
@@ -25,5 +25,3 @@ public class S3NInMemoryFileSystem extends NativeS3FileSystem {
     super(new InMemoryNativeFileSystemStore());
   }
 }
-
-
